% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OEH_COLOURS}
\alias{OEH_COLOURS}
\title{OEH colour definitions}
\format{An object of class \code{character} of length 14.}
\usage{
OEH_COLOURS
}
\description{
A named vector of colour hex codes.
}
\details{
From the OEH Brand Guidelines document: "Our colour palette is divided into
four colour groups - neutrals, sun colours, earth colours and sea colours.
The bright colours allow our charts and graphs to ‘pop’ and have plenty of
contrast. The earth and sea colours provide a gentle reference to our natural
environment."
}
\keyword{datasets}
