% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{oeh_palette}
\alias{oeh_palette}
\title{Returns a function to interpolate an OEH colour palette}
\usage{
oeh_palette(palette = "mixed", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in OEH_PALETTES. One of "mixed", "sun",
"earth", "sea".}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Returns a function to interpolate an OEH colour palette
}
