% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OEH_PALETTES}
\alias{OEH_PALETTES}
\title{Palettes for OEH colours}
\format{An object of class \code{list} of length 4.}
\usage{
OEH_PALETTES
}
\description{
A named list where each element is a vector of OEH colours. Elements are:
\describe{
\item{mixed}{Colours "sun1", "earth1", "sea1"}
\item{sun}{Colours "sun1", "sun2", "sun3"}
\item{earth}{Colours "earth1", "earth2", "earth3"}
\item{sea}{Colours "sea1", "sea2", "sea3"}
\item{neutral}{Colours "neutral1", "neutral2", "neutral3"}
}
}
\keyword{datasets}
