% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_functions.R
\name{scale_colour_oeh}
\alias{scale_colour_oeh}
\title{Colour scale constructor for OEH colours}
\usage{
scale_colour_oeh(palette = "mixed", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in OEH_PALETTES.}

\item{discrete}{Boolean indicating whether the colour aesthetic is discrete or not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to ggplot2::discrete_scale() if discrete is TRUE,
or ggplot2::scale_colour_gradientn() if discrete is FALSE.}
}
\description{
Colour scale constructor for OEH colours
}
\examples{
\dontrun{
ggplot(data = dat, aes(x = group, y = value)) +
  geom_point(aes(colour = subgroup)) +
  scale_colour_oeh(name = "Sub-group", palette = "sun", discrete = TRUE)
}

}
