% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_functions.R
\name{scale_colour_treatment}
\alias{scale_colour_treatment}
\title{Discrete colour scale for plot treatments.}
\usage{
scale_colour_treatment(name = "Plot treatment", breaks = c("control",
  "moderate", "heavy"), values = unname(oeh_colours("black", "earth3",
  "earth1")), ...)
}
\description{
Discrete colour scale for plot treatments.
}
